backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.17.2
aiohttp==3.9.3
httpx==0.26.0
google-generativeai==0.3.2
Pillow==10.2.0
python-dotenv==1.0.0
